Add package example and quickstart
